requirements.txt
amqp==5.2.0
ansicon==1.89.0
arrow==1.3.0
asgiref==3.8.1
billiard==4.2.0
blessed==1.20.0
boto3==1.34.161
botocore==1.34.161
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1
django-environ==0.11.2
django-picklefield==3.2
django-q2==1.6.2
django-storages==1.14.4
django-widget-tweaks==1.5.0
et-xmlfile==1.1.0
jinxed==1.3.0
jmespath==1.0.1
kombu==5.4.0
numpy==2.0.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
plotly==5.23.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
redis==3.5.3
s3transfer==0.10.2
setuptools==72.2.0
six==1.16.0
sqlparse==0.5.1
tenacity==9.0.0
types-python-dateutil==2.9.0.20240316
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
